add cmd.py
