+++ src/multiharp_toolkit/coincidence_counter.py
@@ -131,43 +131,3 @@ def extract(lst: list[int | list[int]]) -> tuple[list[int], list[list[int]]]:
                 s.add(i)
             coincidence_tuples.append(e)
     return (list(s), coincidence_tuples)
-
-
-if __name__ == "__main__":
-    import pyarrow as pa
-    from argparse import ArgumentParser
-
-    parser = ArgumentParser(description="count coincidence from arrow file")
-    parser.add_argument("inputfile", type=str, help="path to the .arrow file")
-    parser.add_argument(
-        "--sync-ch", type=int, help="sync channel", default=0, required=False
-    )
-
-    def parse_channel_and_peak(value):
-        v = value.split(",")
-        return [int(v[0]), float(v[1]), float(v[2])]
-
-    parser.add_argument(
-        "--channel",
-        nargs="+",
-        type=parse_channel_and_peak,
-        help="specify channel and peak window in ps: --channel [ch],[peak_start],[peak_end]. for example: --channel 1,500,600 2,700,800",
-    )
-
-    args = parser.parse_args()
-    print(args)
-    sync_ch = args.sync_ch
-    channels = args.channel
-    counter = CoincidenceCounter(
-        [
-            [ChannelInfo(sync_ch)]
-            + [ChannelInfo(*channels[i]) for i in range(0, len(channels))]
-        ]
-    )
-    counter.process_arrow(args.inputfile)
-    print("coincidence counts:")
-    for k, v in counter.coincidence_counts.items():
-        print(k, v)
-    print("\nch | count")
-    for k, v in counter.number_of_counts.items():
-        print(k, "  ", v)
